pyproject.toml
[project]
name = "cli-whisperer"
version = "0.2.5"
description = "Voice to Text Tool with Smart File Management and OpenAI Formatting"
readme = "README.md"
authors = [
    { name = "VinnyVanGogh", email = "133192356+VinnyVanGogh@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "sounddevice",
    "numpy",
    "pyperclip",
    "scipy",
    "openai",
    "pydantic",
    "python-dotenv",
    "textual",
    "rich",
    "transformers>=4.46.3",
    "torch>=2.5.1",
    "openai-whisper>=20250625",
    "numba>=0.60.0",
]

[project.scripts]
cli-whisperer = "cli_whisperer.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
